Move `Alias` primary key to `Name` (#56)
Issue #, if available:
https://github.com/aws-controllers-k8s/community/issues/1746

Description of changes:
Alias `Name` is the primary key lookup used inside `sdkFind` to determine which of the objects returned by `ListAliases` matches the spec. This pull request moves the `is_primary_key: true` attribute from `TargetKeyId` to `Name`, so that the resource can be adopted by using `alias/<alias-name>` as the `nameOrID` field. 

**Note:** The `nameOrID` field must begin with `alias/` since all alias names returned by the `ListAliases` API are prepended with that prefix.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.